Share Cloud user identity helper
