+++ docs/app/compare/_components/agent-surfaces/cloud-agent-surface.tsx
@@ -1,13 +1,12 @@
 "use client";
 
 import { createCloudChatLLM } from "@/lib/openui-cloud/chat-llm";
-import { CLOUD_USER_ID_HEADER } from "@/lib/openui-cloud/user-id";
+import { CLOUD_USER_ID_HEADER, getOrCreateCloudUserId } from "@/lib/openui-cloud/user-id";
 import { AgentInterface } from "@openuidev/react-ui";
 import { artifactRenderers, chatLibrary, useOpenuiCloudStorage } from "@openuidev/thesys";
 import { useMemo, useState } from "react";
 import type { ComparisonControllerRegistry } from "../comparison-mode-controller";
 import { ComparisonModeControllerBridge } from "../comparison-mode-controller";
-import { getOrCreateCloudUserId } from "./cloud-user-id";
 import { ComparisonSurfaceWelcome } from "./comparison-surface-welcome";
 
 interface CloudAgentSurfaceProps {
